+++ src/Pages/BetaProjectDesktop.jsx
@@ -2,6 +2,7 @@ import { useState, useEffect, useRef } from "react";
 import { Link } from "react-router-dom";
 import { Stack, Box, Alert, AlertTitle, Button } from "@mui/material";
 
+import { useBetaAccount } from "../state/use-beta-account";
 import { useBetaProject } from "../state/use-beta-project";
 import BetaPage from "../components/BetaPage";
 import AddTask from "../components/AddTask";
@@ -9,12 +10,19 @@ import TreeTable from "../TreeTable";
 
 const BetaProject = () => {
   const treeTableRef = useRef();
+  const { error: accountError} = useBetaAccount()
   const { loading, error, uuid, title, data, update } = useBetaProject();
 
   const [src, setSrc] = useState(JSON.stringify(data, null, 2));
   useEffect(() => {
     setSrc(JSON.stringify(data, null, 2));
   }, [data]);
+
+  // This should be removed once we have a real
+  // layout that shield from bad AccountID setup
+  if (accountError) {
+    return null;
+  }
 
   if (loading) {
     return <BetaPage>loading project...</BetaPage>;